% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_structural_analysis.R
\name{analyze_s_subunit_structure}
\alias{analyze_s_subunit_structure}
\title{Analyze S Subunit Structure for Alpha-Helix Spacer}
\usage{
analyze_s_subunit_structure(structure_data, min_helix_len = 15)
}
\arguments{
\item{structure_data}{Data from parse_prostt5_output()}

\item{min_helix_len}{Minimum length to consider as part of the spacer (default 15)}
}
\value{
Data with estimated spacer analysis
}
\description{
Analyzes S subunit structure to find the central alpha-helical spacer
based on long stretches of 'H' (Helix) prediction.

Type I S subunits have characteristic architecture:
\itemize{
\item N-terminal TRD (coil/strand dominant)
\item Central alpha-helical coiled-coil spacer (long H stretches)
\item C-terminal TRD (coil/strand dominant)
}
}
