% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rebase_comparison.R
\name{annotate_with_rebase}
\alias{annotate_with_rebase}
\title{Annotate R-M Candidates with REBASE Results}
\usage{
annotate_with_rebase(rm_data, rebase_results, id_col = "locus_tag")
}
\arguments{
\item{rm_data}{MTase or REase candidate data}

\item{rebase_results}{Results from get_best_rebase_match()}

\item{id_col}{ID column name}
}
\value{
Data with REBASE columns added (best_rebase_match, best_match_identity, rm_type, subunit, rec_seq, etc.)
}
\description{
Adds REBASE match info to MTase/REase data including recognition sequences
}
