% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_operon_analysis.R
\name{calculate_operon_completeness}
\alias{calculate_operon_completeness}
\title{Calculate Operon Completeness}
\usage{
calculate_operon_completeness(operons, rm_genes, id_col)
}
\description{
Calculates completeness based on system type
\itemize{
\item Type I: M + R + S (3 components)
\item Type II: M + R (2 components)
\item Type III: Mod + Res (2 components)
\item Type IV: typically single R component
}
}
\keyword{internal}
