% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mtase_detection.R
\name{classify_mtase_type}
\alias{classify_mtase_type}
\title{Classify Methyltransferase Type}
\usage{
classify_mtase_type(mtase_data, mtase_domains)
}
\arguments{
\item{mtase_data}{Data with MTase hits}

\item{mtase_domains}{Reference PFAM domain table}
}
\value{
Data with added mtase_type column
}
\description{
Internal function to classify MTase type based on PFAM domains
}
\keyword{internal}
