% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_scoring.R
\name{classify_rm_systems}
\alias{classify_rm_systems}
\title{Classify R-M System Type}
\usage{
classify_rm_systems(
  scored_data,
  mtase_data = NULL,
  rease_data = NULL,
  operon_data = NULL
)
}
\arguments{
\item{scored_data}{Scored data from calculate_rm_scores()}

\item{mtase_data}{MTase data with type information}

\item{rease_data}{REase data with type information}

\item{operon_data}{Operon data with predicted type}
}
\value{
Tibble with final classifications
}
\description{
Assigns final R-M system classification based on all evidence
}
