% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_visualization.R
\name{create_evidence_heatmap}
\alias{create_evidence_heatmap}
\title{Create Evidence Heatmap}
\usage{
create_evidence_heatmap(heatmap_data)
}
\description{
Creates the actual heatmap plot
}
\keyword{internal}
