% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mtase_detection.R
\name{detect_mtase_by_keyword}
\alias{detect_mtase_by_keyword}
\title{Detect Methyltransferases by Product Keywords}
\usage{
detect_mtase_by_keyword(dnmb_data, product_col = "product")
}
\arguments{
\item{dnmb_data}{DNMB data frame}

\item{product_col}{Column containing product description}
}
\value{
Tibble with identified methyltransferases
}
\description{
Identifies methyltransferases using product annotation keywords
}
