% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_rease_detection.R
\name{detect_nuclease_by_motif}
\alias{detect_nuclease_by_motif}
\title{Detect Nuclease by Catalytic Motif}
\usage{
detect_nuclease_by_motif(dnmb_data, seq_col = "translation")
}
\arguments{
\item{dnmb_data}{DNMB data frame}

\item{seq_col}{Column containing protein sequences (default: "translation")}
}
\value{
Tibble with catalytic motif hits
}
\description{
Searches protein sequences for known nuclease catalytic motifs
}
