% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_rease_detection.R
\name{detect_rease_by_keyword}
\alias{detect_rease_by_keyword}
\title{Detect Restriction Enzymes by Product Keywords}
\usage{
detect_rease_by_keyword(dnmb_data, product_col = "product")
}
\arguments{
\item{dnmb_data}{DNMB data frame}

\item{product_col}{Product description column}
}
\value{
Tibble with identified REases
}
\description{
Identifies REases using product annotation keywords
}
