% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_rease_detection.R
\name{detect_rease_by_pfam}
\alias{detect_rease_by_pfam}
\title{Detect Restriction Enzymes by PFAM Domain}
\usage{
detect_rease_by_pfam(dnmb_data, pfam_col = NULL)
}
\arguments{
\item{dnmb_data}{DNMB data frame}

\item{pfam_col}{Column containing PFAM annotations}
}
\value{
Tibble with identified REases
}
\description{
Identifies restriction endonucleases using PFAM domains
}
