% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_pipeline.R
\name{export_rm_xlsx}
\alias{export_rm_xlsx}
\title{Export R-M Comprehensive Table to xlsx}
\usage{
export_rm_xlsx(rm_data, output_file, verbose = TRUE)
}
\arguments{
\item{rm_data}{Comprehensive R-M data}

\item{output_file}{Output xlsx file path}

\item{verbose}{Print messages}
}
\value{
Path to exported file
}
\description{
Exports comprehensive table to Excel using xlsx package
}
