% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_structural_analysis.R
\name{export_sequences_for_prostt5}
\alias{export_sequences_for_prostt5}
\title{Export Sequences for ProstT5 Prediction}
\usage{
export_sequences_for_prostt5(s_subunits, output_file, id_col = "locus_tag")
}
\arguments{
\item{s_subunits}{Data frame containing S subunit sequences (must have id_col and 'translation')}

\item{output_file}{Output FASTA path}

\item{id_col}{ID column name}
}
\value{
Path to output file
}
\description{
Exports S subunit sequences to FASTA for ProstT5 prediction
}
