% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rebase_comparison.R
\name{extract_rm_type_from_rebase}
\alias{extract_rm_type_from_rebase}
\title{Extract R-M Type from REBASE Enzyme Name}
\usage{
extract_rm_type_from_rebase(enzyme_name)
}
\arguments{
\item{enzyme_name}{REBASE enzyme name}
}
\value{
Character: Type_I, Type_II, etc.
}
\description{
Determines R-M type from enzyme naming convention
}
