% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_trd_extraction.R
\name{extract_trd_by_keyword}
\alias{extract_trd_by_keyword}
\title{Extract TRD Regions by Keyword}
\usage{
extract_trd_by_keyword(dnmb_data, product_col = "product")
}
\arguments{
\item{dnmb_data}{DNMB data frame}

\item{product_col}{Product column name}
}
\value{
Tibble with TRD-containing genes
}
\description{
Identifies TRD-containing genes by product keywords
}
