% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_trd_extraction.R
\name{extract_trd_by_pfam}
\alias{extract_trd_by_pfam}
\title{Extract TRD Regions by PFAM}
\usage{
extract_trd_by_pfam(dnmb_data, pfam_col = NULL)
}
\arguments{
\item{dnmb_data}{DNMB data frame}

\item{pfam_col}{PFAM column name}
}
\value{
Tibble with TRD-containing genes
}
\description{
Identifies genes with TRD-associated PFAM domains
}
