% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_trd_extraction.R
\name{extract_trd_regions}
\alias{extract_trd_regions}
\title{Combine TRD Extraction Results}
\usage{
extract_trd_regions(
  dnmb_data,
  pfam_col = NULL,
  product_col = "product",
  seq_col = "translation",
  search_motif_sequences = FALSE
)
}
\arguments{
\item{dnmb_data}{DNMB data frame}

\item{pfam_col}{PFAM column}

\item{product_col}{Product column}

\item{seq_col}{Sequence column}

\item{search_motif_sequences}{Whether to search sequence motifs}
}
\value{
Combined tibble with all TRD candidates
}
\description{
Integrates all TRD extraction methods
}
