% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_loading.R
\name{get_column_mapping}
\alias{get_column_mapping}
\title{Get DNMB Column Mapping}
\usage{
get_column_mapping()
}
\value{
Named list of column mappings
}
\description{
Returns standard column name mappings for DNMB data
}
