% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_rease_detection.R
\name{get_nuclease_catalytic_motifs}
\alias{get_nuclease_catalytic_motifs}
\title{Define Catalytic Motifs for Nucleases}
\usage{
get_nuclease_catalytic_motifs()
}
\value{
A tibble with motif patterns and descriptions
}
\description{
Returns regex patterns for known nuclease catalytic motifs
}
