% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_rease_detection.R
\name{get_rease_pfam_domains}
\alias{get_rease_pfam_domains}
\title{Define Restriction Enzyme PFAM Domains}
\usage{
get_rease_pfam_domains(type = "all")
}
\arguments{
\item{type}{RE type: "all", "Type_I", "Type_II", "Type_III", "Type_IV"}
}
\value{
A tibble with PFAM IDs, names, and RE types
}
\description{
Returns PFAM domains associated with restriction endonucleases
}
