% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rebase_comparison.R
\name{get_rebase_cache_dir}
\alias{get_rebase_cache_dir}
\title{Get REBASE Cache Directory}
\usage{
get_rebase_cache_dir()
}
\value{
Path to cache directory
}
\description{
Returns the directory where REBASE data is cached
}
