% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_scoring.R
\name{get_rm_type_keywords}
\alias{get_rm_type_keywords}
\title{Get R-M System Type Keywords}
\usage{
get_rm_type_keywords()
}
\value{
List of keyword patterns by type
}
\description{
Returns keyword patterns for classifying R-M system types
}
