% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_trd_extraction.R
\name{get_trd_pfam_domains}
\alias{get_trd_pfam_domains}
\title{Define TRD-Associated PFAM Domains}
\usage{
get_trd_pfam_domains()
}
\value{
Tibble with TRD-related PFAM information
}
\description{
Returns PFAM domains associated with Target Recognition Domains
}
