% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_trd_extraction.R
\name{get_trd_sequence_motifs}
\alias{get_trd_sequence_motifs}
\title{Define TRD Sequence Motifs}
\usage{
get_trd_sequence_motifs()
}
\value{
Tibble with TRD motif patterns
}
\description{
Returns conserved sequence motifs found in TRD regions
}
