% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_operon_analysis.R
\name{pair_rm_genes}
\alias{pair_rm_genes}
\title{Pair R-M Genes into Operons}
\usage{
pair_rm_genes(rm_genes, all_sorted, id_col, max_intervening, max_distance)
}
\description{
Internal function to pair MTase and REase genes
}
\keyword{internal}
