% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rebase_comparison.R
\name{parse_blast_results}
\alias{parse_blast_results}
\title{Parse BLAST Tabular Output}
\usage{
parse_blast_results(blast_file, verbose = TRUE)
}
\arguments{
\item{blast_file}{BLAST output file}

\item{verbose}{Print diagnostic info}
}
\value{
Tibble with parsed results
}
\description{
Parses BLAST format 6 output with improved enzyme name extraction.
}
