% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_scoring.R
\name{plot_rm_scores}
\alias{plot_rm_scores}
\title{Visualize R-M Scores}
\usage{
plot_rm_scores(scored_data)
}
\arguments{
\item{scored_data}{Scored data}
}
\value{
ggplot object
}
\description{
Creates visualization of R-M system scores
}
