% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_visualization.R
\name{prepare_heatmap_matrix}
\alias{prepare_heatmap_matrix}
\title{Prepare Heatmap Matrix Data}
\usage{
prepare_heatmap_matrix(classified_data, mtase_data, rease_data, max_genes)
}
\description{
Prepares data in long format for heatmap visualization
}
\keyword{internal}
