% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_pipeline.R
\name{print_rmscan_summary}
\alias{print_rmscan_summary}
\title{Print RMscan Results Summary}
\usage{
print_rmscan_summary(results)
}
\arguments{
\item{results}{Results from rmscan_pipeline()}
}
\description{
Prints formatted summary of analysis results
}
