% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rebase_comparison.R
\name{run_blastp_rebase}
\alias{run_blastp_rebase}
\title{Run Local BLASTP Against REBASE}
\usage{
run_blastp_rebase(
  query_fasta,
  rebase_db,
  output_file = NULL,
  evalue = 1e-05,
  num_threads = 4,
  max_target_seqs = 5
)
}
\arguments{
\item{query_fasta}{Path to query FASTA file}

\item{rebase_db}{Path to REBASE BLAST database (or FASTA to create db)}

\item{output_file}{Output file path}

\item{evalue}{E-value threshold}

\item{num_threads}{Number of threads}

\item{max_target_seqs}{Maximum hits per query}
}
\value{
Path to BLAST results file
}
\description{
Runs blastp against REBASE database
}
