% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rebase_comparison.R
\name{run_hmmscan_pfam}
\alias{run_hmmscan_pfam}
\title{Run HMMER Search Against PFAM (Alternative to BLAST)}
\usage{
run_hmmscan_pfam(
  query_fasta,
  pfam_db,
  output_file = NULL,
  evalue = 1e-05,
  num_threads = 4
)
}
\arguments{
\item{query_fasta}{Path to query FASTA file}

\item{pfam_db}{Path to PFAM-A.hmm database}

\item{output_file}{Output file path}

\item{evalue}{E-value threshold}

\item{num_threads}{Number of threads}
}
\value{
Path to HMMER results file or NULL if not available
}
\description{
Uses hmmscan against PFAM for more sensitive R-M detection.
This is an alternative/complement to BLAST when detecting divergent R-M systems.

Note: This requires PFAM database to be installed locally.
Download from: ftp://ftp.ebi.ac.uk/pub/databases/Pfam/current_release/
}
