% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_trd_extraction.R
\name{search_trd_motifs}
\alias{search_trd_motifs}
\title{Search TRD Sequence Motifs}
\usage{
search_trd_motifs(dnmb_data, seq_col = "translation")
}
\arguments{
\item{dnmb_data}{DNMB data with sequences}

\item{seq_col}{Sequence column name}
}
\value{
Tibble with motif search results
}
\description{
Searches protein sequences for conserved TRD motifs
}
